refactor(cli): consolidate toml patcher tests; extract oauth-format helper
- Move formatOAuthAuthMessage from login.ts to src/cli/utils/oauth-format.ts.
- login.ts and onboard.ts now import from the new shared location (login.ts also
  re-exports it until Task 14 deletes login.ts).
- Move all OAuth patcher tests (addProviderToOAuthTier, removeProviderFromOAuthTier,
  setProviderAsPrimary, plus the 'preserve adjacent commented templates' regression
  suite) from login.test.ts into toml-patchers.test.ts so they live next to the
  implementation.
- Move the formatOAuthAuthMessage tests into a new oauth-format.test.ts.
- Delete the now-empty login.test.ts.
